Make sort logic available outside sort-by
