+++ crates/nu-command/src/lib.rs
@@ -19,6 +19,7 @@ mod path;
 mod platform;
 mod random;
 mod shells;
+mod sort_utils;
 mod strings;
 mod system;
 mod viewers;
@@ -45,6 +46,7 @@ pub use path::*;
 pub use platform::*;
 pub use random::*;
 pub use shells::*;
+pub use sort_utils::*;
 pub use strings::*;
 pub use system::*;
 pub use viewers::*;
